// src/app/shared/model/produit-simple/tarif-unit-var.model.ts
export class TarifUnitVar {
  _id?: string;
  tarifUHT?: number;
  monnaie?: string;
  appToAllDistrs?: boolean;
  concernDistrs?: string[];
  noConcernDistrs?: string[];
  appToAllZones?: boolean;
  concernZones?: string[];
  noConcernZones?: string[];
  description?: string;
  commentaire?: string;
  produitService?: any;
}